requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
greenlet==3.2.4
idna==3.10
numpy==2.2.6
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
